Vary output according to condition

An R equivalent of the SQL CASE WHEN statement, if no cases match, NA is returned. This function is supported by data.table::fcase (a wrapper), but uses another syntax as in dplyr::case_when.

case_when_dt(...)

Arguments

...
A sequence of two-sided formulas. The left hand side (LHS) determines which values match this case. The right hand side (RHS) provides the replacement value. The LHS must evaluate to a logical vector. The RHS does not need to be logical, but all RHSs must evaluate to the same type of vector.

Value

A vector

Details

Supported by fcase, case_when_dt is supposed to be faster than dplyr::case_when. The only difference is the separator of the function (tidyfst::case_when_dt) uses "~" as in dplyr::case_when, while data.table::fcase uses ",". The data.table::fcase is also exported in tidyfst, if the expressions include one or more "~" in either side, please use fcase instead of case_when_dt.

See also

Examples


x = 1:10
res1 = fcase(
  x < 5L, 1L,
  x > 5L, 3L
)
res2 = case_when_dt(
  x < 5 ~ 1,
  x > 5 ~ 3
)

#> Error in fcase(x < 5, 1, x > 5, 3): object 'x' not found

setequal(res1,res2)

#> Error in as.vector(y): object 'res2' not found


# if you want to use "~" in the expressions, you can only use `fcase`
## Correct
fcase(
  x < 5L, "~",
  x > 5L, "~~"
)

#>  [1] "~"  "~"  "~"  "~"  NA   "~~" "~~" "~~" "~~" "~~"

## Wrong
case_when_dt(
  x < 5 ~ "~",
  x > 5 ~ "~~"
)

#> Error in fcase(x < 5, ",", x > 5, ",,"): object 'x' not found