Objects exported from other packages

These objects are imported from other packages. Follow the links below to see their documentation.

data.table

%chin%, %like%, as.data.table, between, CJ, copy, data.table, fcase, fcoalesce, fintersect, fread, frollapply, fsetdiff, fsetequal, funion, fwrite, key, like, rbindlist, rleid, rleidv, set, setDF, setDT, setnames, tables, uniqueN

stringr

%>%