Pivot data from long to wide

Transform a data frame from long format to wide by increasing the number of columns and decreasing the number of rows.

wider_dt(.data, ..., name, value = NULL, fun = NULL, fill = NA)

Arguments

.data
A data.frame

...
Optional. The unchanged group in the transformation. Could use integer vector, could receive what select_dt receives.

name
Chracter.One column name of class to spread

value
Chracter.One column name of value to spread. If NULL, use all other variables.

fun
Should the data be aggregated before casting? Defaults to NULL, which uses length for aggregation. If a function is provided, with aggregated by this function.

fill
Value with which to fill missing cells. Default uses NA.

Value

data.table

Details

The parameter of `name` and `value` should always be provided and should be explicit called (with the parameter names attached).

See also

Examples

 stocks = data.frame(
   time = as.Date('2009-01-01') + 0:9,
   X = rnorm(10, 0, 1),
   Y = rnorm(10, 0, 2),
   Z = rnorm(10, 0, 4)
 ) %>%
   longer_dt(time) -> longer_stocks

 longer_stocks

#>           time   name      value
#>         <Date> <fctr>      <num>
#>  1: 2009-01-01      X  1.8150308
#>  2: 2009-01-02      X -0.9369726
#>  3: 2009-01-03      X  0.8613005
#>  4: 2009-01-04      X  0.6216399
#>  5: 2009-01-05      X -2.1443322
#>  6: 2009-01-06      X  0.4847217
#>  7: 2009-01-07      X -0.4472743
#>  8: 2009-01-08      X  0.5460971
#>  9: 2009-01-09      X  1.4441867
#> 10: 2009-01-10      X  0.2428887
#> 11: 2009-01-01      Y  0.3370242
#> 12: 2009-01-02      Y  0.4781293
#> 13: 2009-01-03      Y  0.4726428
#> 14: 2009-01-04      Y -0.5182383
#> 15: 2009-01-05      Y  1.2980914
#> 16: 2009-01-06      Y -2.4352820
#> 17: 2009-01-07      Y  1.6839408
#> 18: 2009-01-08      Y -3.2384213
#> 19: 2009-01-09      Y  0.5283319
#> 20: 2009-01-10      Y  0.1946214
#> 21: 2009-01-01      Z  7.1554614
#> 22: 2009-01-02      Z -3.9547311
#> 23: 2009-01-03      Z  7.5636128
#> 24: 2009-01-04      Z  4.8033099
#> 25: 2009-01-05      Z -7.2528441
#> 26: 2009-01-06      Z  1.4817289
#> 27: 2009-01-07      Z -1.9307071
#> 28: 2009-01-08      Z  0.4051379
#> 29: 2009-01-09      Z  0.1563092
#> 30: 2009-01-10      Z  4.2593028
#>           time   name      value


 longer_stocks %>%
   wider_dt("time",
            name = "name",
            value = "value")

#> Key: <time>
#>           time          X          Y          Z
#>         <Date>      <num>      <num>      <num>
#>  1: 2009-01-01  1.8150308  0.3370242  7.1554614
#>  2: 2009-01-02 -0.9369726  0.4781293 -3.9547311
#>  3: 2009-01-03  0.8613005  0.4726428  7.5636128
#>  4: 2009-01-04  0.6216399 -0.5182383  4.8033099
#>  5: 2009-01-05 -2.1443322  1.2980914 -7.2528441
#>  6: 2009-01-06  0.4847217 -2.4352820  1.4817289
#>  7: 2009-01-07 -0.4472743  1.6839408 -1.9307071
#>  8: 2009-01-08  0.5460971 -3.2384213  0.4051379
#>  9: 2009-01-09  1.4441867  0.5283319  0.1563092
#> 10: 2009-01-10  0.2428887  0.1946214  4.2593028


 longer_stocks %>%
   mutate_dt(one = 1) %>%
   wider_dt("time",
            name = "name",
            value = "one")

#> Key: <time>
#>           time     X     Y     Z
#>         <Date> <num> <num> <num>
#>  1: 2009-01-01     1     1     1
#>  2: 2009-01-02     1     1     1
#>  3: 2009-01-03     1     1     1
#>  4: 2009-01-04     1     1     1
#>  5: 2009-01-05     1     1     1
#>  6: 2009-01-06     1     1     1
#>  7: 2009-01-07     1     1     1
#>  8: 2009-01-08     1     1     1
#>  9: 2009-01-09     1     1     1
#> 10: 2009-01-10     1     1     1


## using "fun" parameter for aggregation
DT <- data.table(v1 = rep(1:2, each = 6),
                 v2 = rep(rep(1:3, 2), each = 2),
                 v3 = rep(1:2, 6),
                 v4 = rnorm(6))
## for each combination of (v1, v2), add up all values of v4
DT %>%
  wider_dt(v1,v2,
           value = "v4",
           name = ".",
           fun = sum)

#> Key: <v1, v2>
#>       v1    v2          .
#>    <int> <int>      <num>
#> 1:     1     1 -2.1096711
#> 2:     1     2 -1.4708456
#> 3:     1     3 -0.1688983
#> 4:     2     1 -2.1096711
#> 5:     2     2 -1.4708456
#> 6:     2     3 -0.1688983